requirements.txt
pycryptodome==3.20.0
web3==6.19.0